Rename: 폴더이동 및 불필요한 내용제거
