+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "portfolio App",
+  "name": "portfolio React App",
   "icons": [],
   "start_url": ".",
   "display": "standalone",
